Fix critical math and memory bugs in PG-MambaGAN
- Fix Mathematical Explosion in NPS Loss: Applied `torch.log1p` to power spectrum and removed the DC component to prevent exploding gradients.
- Fix AMP / FFT Instability: Moved the frequency loss evaluation outside the `torch.amp.autocast` block and cast inputs to float.
- Implement Strict HU to [-1, 1] Normalization: Added strict `[-1000, 1000]` clipping and mathematical mapping to `[-1.0, 1.0]` for dataset inputs.
- Fix VRAM Memory Leak: Removed redundant `.copy()` calls after augmentations and wrapped arrays in `np.ascontiguousarray()` to prevent negative stride errors.

+++ training/trainer_pt.py
@@ -279,11 +279,11 @@ class Trainer:
                     total_g_loss = total_g_loss + loss_perc
                     g_losses["g_perc"] = g_losses.get("g_perc", 0) + loss_perc.item() / self.grad_accum
                 
-                # Frequency
-                if "freq" in self.g_loss_fns:
-                    loss_freq = self.g_loss_fns["freq"](fake, ndct) * self.loss_weights["freq"]
-                    total_g_loss = total_g_loss + loss_freq
-                    g_losses["g_freq"] = g_losses.get("g_freq", 0) + loss_freq.item() / self.grad_accum
+            # Frequency (outside AMP — uses FFT which may be unstable)
+            if "freq" in self.g_loss_fns:
+                loss_freq = self.g_loss_fns["freq"](fake.float(), ndct.float()) * self.loss_weights["freq"]
+                total_g_loss = total_g_loss + loss_freq
+                g_losses["g_freq"] = g_losses.get("g_freq", 0) + loss_freq.item() / self.grad_accum
             
             # NPS loss (outside AMP — uses FFT which may not support bfloat16)
             if "nps" in self.g_loss_fns:
